done with day 3

+++ solutions/day003.py
@@ -1,7 +1,7 @@
 # Erhmprez
 #Area of a triangle.
 
-base_triangle = float(input('Enter the base of the triangle: '))
-height_triangle = float(input('Enter the height of the triangle: '))
+base_triangle = int(input('Enter the base of the triangle: '))
+height_triangle = int(input('Enter the height of the triangle: '))
 area = 0.5 * base_triangle * height_triangle
 print('The area of the Triangle is ', area)
